Revert "[DD-269] Node.JS & Python Guided Tours (#6111)"
This reverts commit af3dc86fc708c648bbadd9d9661cb2d4bd14c42c.
