+++ jekyll/_cci2/index.md
@@ -12,33 +12,6 @@ Use the tutorials, samples, how-to, and reference documentation to learn CircleC
 <hr class="hidden-xs" />
 
 <div class="row">
-  <div class="treatment col-xs-12">
-    <span id="homepage-guide-links"><h2>Examples and Guides</h2><img src="{{ site.baseurl }}/assets/img/compass/new.svg" alt="New"></span>
-    <p>Get started quickly: follow step-by-step <a href="{{site.baseurl}}/2.0/tutorials/">guides</a> or explore a sample app.</p>
-  </div>
-  <div class="treatment col-xs-12 col-sm-6">
-    <a class="no-external-icon col-sm-12" href="{{site.baseurl}}/2.0/language-javascript/">
-      <div class="card col-sm-12">
-        {% capture node-js-card %}
-          {% include snippets/language-card.md lang="Node.JS" anchor="true" %}
-        {% endcapture %}
-        {{ node-js-card | markdownify }}
-      </div>
-    </a>
-  </div>
-  <div class="treatment col-xs-12 col-sm-6">
-    <a class="no-external-icon col-sm-12" href="{{site.baseurl}}/2.0/language-python/">
-      <div class="card col-sm-12">
-        {% capture python-card %}
-          {% include snippets/language-card.md lang="Python" anchor="true" %}
-        {% endcapture %}
-        {{ python-card | markdownify }}
-      </div>
-    </a>
-  </div>
-  <div class="treatment col-xs-12">
-    <hr />
-  </div>
   <div class="col-xs-12 col-sm-6">
     <h2>Get Started</h2>
     <p>Get started with CircleCI automated builds.</p>
