+++ jekyll/_config.yml
@@ -7,14 +7,8 @@ description: > # this means to ignore newlines until "baseurl:"
 
 url: "https://ui.circleci.com" # the base hostname & protocol for your site
 baseurl: "/docs" # the subpath of your site, e.g. /blog
-gh_help_articles_url: "https://help.github.com/articles"
 gh_url: "https://github.com/circleci/circleci-docs"
-gh_private_org_url: "https://github.com/circleci"
-gh_public_org_url: "https://github.com/CircleCI-Public"
-cci_public_org_url: "https://app.circleci.com/pipelines/github/CircleCI-Public"
 devhub_base_url: "https://circleci.com/developer"
-blog_base_url: "https://circleci.com/blog"
-support_base_url: "https://support.circleci.com"
 noindex: false
 
 exclude:
