+++ jekyll/_includes/doc-footer.html
@@ -1,7 +1,4 @@
-{% if page.layout == "classic-docs-experimental" or page.page-type == "homepage" %}
-	{% assign class = "loading-deferred" %}
-{% endif %}
-<div class="{{class}}">
+<div>
 	{% include_localized edit-on-github.html site=site page=page %}
 	{% include_localized license-notice.html %}
 </div>
