+++ jekyll/_includes/global-footer.html
@@ -1,7 +1,4 @@
-{% if page.layout == "classic-docs-experimental" or page.page-type == "homepage" %}
-  {% assign class = "loading-deferred" %}
-{% endif %}
-<footer class="docs-footer inverted {{class}}" data-analytics="clicked-footer" data-analytics-parent>
+<footer class="docs-footer inverted" data-analytics="clicked-footer" data-analytics-parent>
   <div class="blog-container">
     <div class="nav-row">
       <a href="{{ '/' | relative_url }}" class="lockup-link" >
