--- jekyll/_includes/no-script.html
@@ -1,8 +1,0 @@
-<!--
-    Since by default pages in experiment display none both control and treatment
-    content. It usually depends on JS to display the correct content but in the
-    case of no JavaScript, this noscript tag defaults to showing control content.
---->
-<noscript>
-    <style>.row,.control{display:initial}.loading-deferred{display:block}#full-height{visibility:visible}</style>
-</noscript>
