--- jekyll/_layouts/classic-docs-experimental.html
@@ -1,16 +1,0 @@
----
-layout: classic-docs
-section: cci2
----
-<div class="control {{ page.categories | join: ' ' }}">
-    {{content}}
-</div>
-<div class="treatment no_toc_section {{ page.categories | join: ' ' }}">
-    {% if page.categories contains "language-guides" %}
-        {% capture lang_guide %}
-            {% include_relative language-{{page.short-title | downcase}}-experimental.md %}
-        {% endcapture %}
-{{ lang_guide | split: "---" | last | markdownify }}
-    {% endif %}
-</div>
-<style>#full-height{visibility:hidden}</style>
