+++ src/js/app.js
@@ -14,7 +14,7 @@ window.Cookie = Cookie;
 window.AnalyticsClient = services.AnalyticsClient; // because it only has static methods, AnalyticsClient is not instantiated
 window.OptimizelyClient = new services.OptimizelyClient();
 
-import site from './site';
+import './site';
 
 services.lang.init();
 services.rum.init();
@@ -25,25 +25,5 @@ $(() => {
   services.instantsearch.init();
   services.highlightjsBadge.init();
   services.progressbar.init();
-
-  // Boolean whether or not URL is in Guided Tour experiment
-  const isGatedPath =
-    ['/docs/', '/docs/2.0/'].includes(window.location.pathname) ||
-    ['-preview/', 'view/2.0/'].includes(window.location.pathname.slice?.(-9)) ||
-    /language-(javascript|python)/gm.test(window.location.pathname);
-  if (!isGatedPath) {
-    /** If URL is not in Guided Tour experiment, then we can attach eventListeners
-     *  for highlighting the Table of Contents in the sidebar as user scrolls.
-     *
-     *  If URL is in Guided Tour experiment, then it must wait for experiment
-     *  logic to complete before attaching eventListeners since the ToC is
-     *  different in treatment vs control.
-     */
-    site.toc.highlightTocOnScrollOnce();
-  }
-
-  import(/* webpackPrefetch: true */ './experiments').then(
-    ({ default: { languageGuides } = {} }) =>
-      isGatedPath ? languageGuides() : null,
-  ); // ensure languageGuides is loaded; // imports all experiments
+  import(/* webpackPrefetch: true */ './experiments'); // imports all experiments
 });
