+++ src/js/experiments/index.js
@@ -1,10 +1,8 @@
 import forceAll from './forceAll';
 import videoTutorials from './videoTutorials';
-import languageGuides from './languageGuides';
 import './snippetFeedback';
 
 export default {
   forceAll,
-  languageGuides,
   videoTutorials,
 };
