+++ src/js/site/main.js
@@ -3,10 +3,8 @@ import { highlightURLHash } from './highlightURLHash';
 
 hljs.initHighlightingOnLoad();
 
-const SHOW_EVENTS = ['mouseover', 'hover', 'mouseenter', 'focus'];
-const HIDE_EVENTS = ['mouseout', 'mouseleave', 'blur'];
-const HEADER_TAGS =
-  'article h1, article h2, article h3, article h4, article h5, article h6';
+const showEvents = ['mouseover', 'hover', 'mouseenter', 'focus'];
+const hideEvents = ['mouseout', 'mouseleave', 'blur'];
 
 // compiles an object of parameters relevant for analytics event tracking.
 // takes an optional DOM element and uses additional information if present.
@@ -119,11 +117,11 @@ function renderVersionBlockPopover() {
       destroy();
     }
 
-    SHOW_EVENTS.forEach((event) => {
+    showEvents.forEach((event) => {
       badge.addEventListener(event, show);
     });
 
-    HIDE_EVENTS.forEach((event) => {
+    hideEvents.forEach((event) => {
       badge.addEventListener(event, hide);
     });
   });
@@ -257,8 +255,7 @@ $(document).ready(function () {
   var tooltip = document.querySelector('.tooltip-popover');
 
   // Give article headings direct links to anchors
-  $(HEADER_TAGS)
-    .not(HEADER_TAGS.replace(/article/gm, '.card'))
+  $('article h1, article h2, article h3, article h4, article h5, article h6')
     .filter('[id]')
     .each(function () {
       var isMainTitle = $(this).prop('nodeName') === 'H1';
@@ -354,11 +351,11 @@ $(document).ready(function () {
       document.querySelectorAll('.fa-link').forEach((icon) => {
         makePopper(icon);
 
-        SHOW_EVENTS.forEach((event) => {
+        showEvents.forEach((event) => {
           icon.parentElement.addEventListener(event, icon.show);
         });
 
-        HIDE_EVENTS.forEach((event) => {
+        hideEvents.forEach((event) => {
           icon.parentElement.addEventListener(event, icon.hide);
         });
 
@@ -369,7 +366,7 @@ $(document).ready(function () {
     }
   });
 
-  $(HEADER_TAGS)
+  $('article h1, article h2, article h3, article h4, article h5, article h6')
     .filter('[id]')
     .hover(function () {
       $(this).find('i').toggle();
@@ -407,14 +404,14 @@ $(function () {
   const timeposted = document.getElementById('time-posted-on');
   let popperInstance = null;
 
-  SHOW_EVENTS.forEach((event) => {
+  showEvents.forEach((event) => {
     timeposted?.addEventListener(event, () => {
       tooltiptime.setAttribute('data-show', '');
       popperInstance = createPopper(timeposted, tooltiptime, {});
     });
   });
 
-  HIDE_EVENTS.forEach((event) => {
+  hideEvents.forEach((event) => {
     timeposted?.addEventListener(event, () => {
       tooltiptime.removeAttribute('data-show');
       if (popperInstance) {
