--- src/styles/_language-guide.scss
@@ -1,188 +1,0 @@
-.row, .control, .treatment, .loading-deferred {
-    display: none;
-
-    a.no-external-icon.col-sm-12 {
-        padding: 0;
-
-        &:hover {
-            div.card.col-sm-12 {
-                background: #F7F7F7;
-            }
-        }
-    }
-
-    .treatment.col-xs-12.col-sm-6 {
-        margin-top: 24px;
-    }
-}
-
-#current-variation {
-    display: initial;
-}
-
-.card {
-    img{
-        margin: 15px 0 0 0;
-    }
-
-    h3 {
-        line-height: 31.2px;
-    }
-
-    h1, h2, h3, h4, h5, h6 {
-        margin-top: 12px;
-        margin-bottom: 4px;
-    }
-
-    h3 + p {
-        margin-bottom: 4px;
-        line-height: 24px;
-    }
-
-    h3 + p + p {
-        margin-bottom: 16px;
-
-        strong {
-            font-weight: 500;
-        }
-    }
-}
-
-#cards {
-    display: flex;
-    margin-top: 25px;
-
-    .card {
-        width: 50%;
-        padding: 24px;
-        color: #555555;
-
-        img {
-            margin: 0;
-        }
-
-        h1, h2, h3, h4, h5, h6 {
-            margin-top: 16px;
-            color: black;
-        }
-
-        h3 + p {
-            margin-bottom: 12px;
-        }
-
-        & > ul {
-            padding: 0;
-        }
-
-        #completion-time {
-            padding: 0;
-            font-size: 15px;
-            width: fit-content;
-            line-height: 18px;
-
-            * {
-                padding: 0;
-                margin: 0;
-            }
-
-            p {
-                padding: 2px 2px 1px 2px;
-            }
-
-            img {
-                padding: 7px 0 9px 15px;
-                width: 39px;
-            }
-
-            span {
-                padding: 0 19px 0 4.5px;
-                line-height: 18px;
-            }
-        }
-
-        ul {
-            display: list-item;
-            list-style-type: none;
-            line-height: 28.8px;
-
-            &#markdown-toc {
-                display: initial;
-                list-style-type: none;
-            }
-        }
-
-        @media(min-width: $screen-sm) {
-            &:not(#completion-time):first-of-type {
-                border-top-right-radius: 0;
-                border-bottom-right-radius: 0;
-            }
-
-            &:not(#completion-time):last-of-type {
-                border-top-left-radius: 0;
-                border-bottom-left-radius: 0;
-                border-left: 0;
-            }
-        }
-    }
-
-    #card-or-card-container {
-        position: relative;
-        width: 0;
-        height: 100%;
-
-        #card-or-card {
-            position: absolute;
-            font-size: 12px;
-            top: 26.5em;
-            right: -15.5px;
-            background: #F7F7F7;
-            border-radius: 75px;
-            font-weight: 700;
-            padding: 7px 8px 7px 8px;
-            z-index: 10;
-        }
-    }
-
-    @media(max-width: $screen-sm) {
-        flex-direction: column; 
-
-        .card {
-            width: 100%;
-
-            &:not(#completion-time):first-of-type {
-                border-bottom-left-radius: 0;
-                border-bottom-right-radius: 0;
-                margin-bottom: 0;
-            }
-
-            &:not(#completion-time):last-of-type {
-                border-top-left-radius: 0;
-                border-top-right-radius: 0;
-                border-top: 0;
-            }
-        }
-
-        #card-or-card-container {
-            width: 100%;
-            height: 0;
-
-            #card-or-card {
-                top: 0;
-                right: 0;
-                margin: -16.5px calc((100% - 42px) / 2)
-            }
-        }
-    }
-}
-
-#homepage-guide-links {
-    display: flex;
-
-    img {
-        padding: 11px 0 0 8px;
-    }
-
-    & + p {
-        margin: 0;
-    }
-}
