+++ src/styles/main.scss
@@ -16,7 +16,6 @@
 @import 'snippets';
 @import 'utilities';
 @import 'progressbar';
-@import 'language-guide';
 
 // General
 body {
